Cast wave's PE to int16 and kmall's PE to int8

+++ scripts/histograms.py
@@ -21,7 +21,7 @@ def waveGenerator(file):
     # Find PE using a bash pipeline (faster)
     wave_pe = subprocess.run(["wave/wave_pe.sh", file], capture_output=True)
     # Dataframe for PE
-    df_pe = pd.read_csv(StringIO(wave_pe.stdout.decode('utf-8')), header=None, names=["PE"], dtype="int")
+    df_pe = pd.read_csv(StringIO(wave_pe.stdout.decode('utf-8')), header=None, names=["PE"], dtype="int16")
     # Dataframe for original samples
     df_samples = pd.DataFrame(np.fromfile(file, dtype="int16"), columns=["samples"], dtype="int16")
     # Return both dataframes
@@ -31,7 +31,7 @@ def kmallGenerator(file):
     # Find PE using a bash pipeline (faster)
     kmall_pe = subprocess.run(["kmall/kmall_pe.sh", file], capture_output=True)
     # Dataframe for PE
-    df_pe = pd.read_csv(StringIO(kmall_pe.stdout.decode('utf-8')), header=None, names=["PE"], dtype="int")
+    df_pe = pd.read_csv(StringIO(kmall_pe.stdout.decode('utf-8')), header=None, names=["PE"], dtype="int8")
     # Dataframe for original samples
     df_samples = pd.DataFrame(np.fromfile(file, dtype="int8"), columns=["samples"], dtype="int8")
     # Return both dataframes
